Rename file: dirwalker -> dir_walker
Felt like this file was missing an '_'
